pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-gemini-hub"
version = "1.0.0"
description = "The ultimate Gemini Hub for ComfyUI: Text-to-Speech, Speech-to-Text, Chat, and all Google AI capabilities in one place"
authors = [{name = "Makhmudjon Inadullaev", email = "makhmudjon.dev@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
keywords = ["comfyui", "gemini", "hub", "collection", "tts", "stt", "chat", "google", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "google-genai",
    "torch",
    "torchaudio",
]

[project.urls]
Homepage = "https://github.com/makhmudjon-inadullaev/comfyui-gemini-hub"
Repository = "https://github.com/makhmudjon-inadullaev/comfyui-gemini-hub"
Issues = "https://github.com/makhmudjon-inadullaev/comfyui-gemini-hub/issues"
